requirements.txt
Django==3.2.9
djangorestframework==3.12.4
pillow==8.1.2
django-cors-headers==3.7.0
gunicorn==20.1.0